refractor: Code connecting data to Adapter relocated
Within the MainActivity, the code that connects the Weather Data to the
WeatherAdapter has been put back into the NetworkListener.

It was originally like this, however it was moved to a separate
method to verify if it was the cause of a crash.

It was not the cause of the crash, therefore it is being returned to
ensure in order to remove an unnecessary method.